Workbench shutdown proper handle (fuck crazy juno developers)
Former-commit-id: 61dba82195ee4c000973acd8044d5a062d50d7aa

+++ branches/juno-migration/plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/ui/UIUtils.java
@@ -1040,6 +1040,10 @@ public class UIUtils {
 
     public static void removeFocusTracker(IServiceLocator serviceLocator, Control control)
     {
+        if (DBeaverCore.getInstance().getWorkbench().isClosing()) {
+            // TODO: it is a bug in eclipse. During workbench shutdown disposed service returned.
+            return;
+        }
         final IFocusService focusService = (IFocusService) serviceLocator.getService(IFocusService.class);
         if (focusService != null) {
             focusService.removeFocusTracker(control);
